Completed refactor before fixes
