+++ pyproject.toml
@@ -71,3 +71,4 @@ ubc-solar-physics = "^1.1.0"
 pydantic = "^2.10.6"
 anytree = "^2.12.1"
 graphviz = "^0.20.3"
+pyowm = "^3.3.0"
