+++ simulation/config/settings_FSGP.toml
@@ -24,7 +24,6 @@ weather_period = "Daily"
 
 [competition_config.route_config]
  coordinates = [
-    [38.9281815, -95.677021],
     [ 36.99932082, -86.37230251],
     [ 36.99940337, -86.37214797],
     [ 36.99948814, -86.37199725],
@@ -314,6 +313,4 @@ weather_period = "Daily"
     [ 36.99916641, -86.37284378],
     [ 36.99921127, -86.37261758],
     [ 36.99924349, -86.37245211],
-    [ 36.99932082, -86.37230251],
-    [38.9282115, -95.6770268]
 ]
