+++ simulation/model/__init__.py
@@ -1,5 +1,6 @@
 from .Simulation import Simulation
 from .Model import Model
+from .ModelBuilder import ModelBuilder
 
 __all__ = [
     "Simulation",
